Cache all attribute names with custom casts for improved performance.
Motivation: I have a class where just one attribute is custom-cast. Every single access to any of the other attributes would call `getCustomCasts`, which in turn invokes the autoloader in the `is_subclass_of` call for every single attribute in the `casts` array.

Caching the result of `$this->getCustomCasts()` has doubled performance for me in some cases.